+++ internal/internal_workflow_client_test.go
@@ -2903,6 +2903,37 @@ func TestGetWorkflowStartRequest(t *testing.T) {
 			},
 		},
 		{
+			name: "with RequestID",
+			options: StartWorkflowOptions{
+				ID:                              workflowID,
+				RequestID:                       uuid.Parse("550e8400-e29b-41d4-a716-446655440000"),
+				TaskList:                        tasklist,
+				ExecutionStartToCloseTimeout:    10 * time.Second,
+				DecisionTaskStartToCloseTimeout: 5 * time.Second,
+			},
+			workflowFunc: func(ctx Context) {},
+			wantRequest: &shared.StartWorkflowExecutionRequest{
+				Domain:     common.StringPtr(domain),
+				RequestId:  common.StringPtr("550e8400-e29b-41d4-a716-446655440000"),
+				WorkflowId: common.StringPtr(workflowID),
+				WorkflowType: &shared.WorkflowType{
+					Name: common.StringPtr("go.uber.org/cadence/internal.TestGetWorkflowStartRequest.func2"),
+				},
+				TaskList: &shared.TaskList{
+					Name: common.StringPtr(tasklist),
+				},
+				ExecutionStartToCloseTimeoutSeconds: common.Int32Ptr(10),
+				TaskStartToCloseTimeoutSeconds:      common.Int32Ptr(5),
+				DelayStartSeconds:                   common.Int32Ptr(0),
+				JitterStartSeconds:                  common.Int32Ptr(0),
+				FirstRunAtTimestamp:                 common.Int64Ptr(0),
+				CronSchedule:                        common.StringPtr(""),
+				Header:                              &shared.Header{Fields: map[string][]byte{}},
+				WorkflowIdReusePolicy:               shared.WorkflowIdReusePolicyAllowDuplicateFailedOnly.Ptr(),
+				CronOverlapPolicy:                   shared.CronOverlapPolicySkipped.Ptr(),
+			},
+		},
+		{
 			name: "missing TaskList",
 			options: StartWorkflowOptions{
 				ID:                              workflowID,
@@ -3018,7 +3049,9 @@ func TestGetWorkflowStartRequest(t *testing.T) {
 
 			// set the randomized fields in the expected request before comparison
 			tc.wantRequest.Identity = &wc.identity
-			tc.wantRequest.RequestId = gotReq.RequestId
+			if tc.wantRequest.RequestId == nil {
+				tc.wantRequest.RequestId = gotReq.RequestId
+			}
 
 			assert.Equal(t, tc.wantRequest, gotReq)
 		})
@@ -3037,6 +3070,41 @@ func TestGetSignalWithStartRequest(t *testing.T) {
 		wantRequest  *shared.SignalWithStartWorkflowExecutionRequest
 		wantErr      string
 	}{
+		{
+			name:       "with RequestID",
+			workflowID: workflowID,
+			signalName: "signal",
+			options: StartWorkflowOptions{
+				ID:                              workflowID,
+				RequestID:                       uuid.Parse("550e8400-e29b-41d4-a716-446655440000"),
+				TaskList:                        tasklist,
+				ExecutionStartToCloseTimeout:    10 * time.Second,
+				DecisionTaskStartToCloseTimeout: 5 * time.Second,
+			},
+			workflowFunc: func(ctx Context) {},
+			wantRequest: &shared.SignalWithStartWorkflowExecutionRequest{
+				Domain:     common.StringPtr(domain),
+				RequestId:  common.StringPtr("550e8400-e29b-41d4-a716-446655440000"),
+				WorkflowId: common.StringPtr(workflowID),
+				WorkflowType: &shared.WorkflowType{
+					Name: common.StringPtr("go.uber.org/cadence/internal.TestGetSignalWithStartRequest.func1"),
+				},
+				TaskList: &shared.TaskList{
+					Name: common.StringPtr(tasklist),
+				},
+				ExecutionStartToCloseTimeoutSeconds: common.Int32Ptr(10),
+				TaskStartToCloseTimeoutSeconds:      common.Int32Ptr(5),
+				SignalName:                          common.StringPtr("signal"),
+				SignalInput:                         []byte("null\n"),
+				DelayStartSeconds:                   common.Int32Ptr(0),
+				JitterStartSeconds:                  common.Int32Ptr(0),
+				FirstRunAtTimestamp:                 common.Int64Ptr(0),
+				CronSchedule:                        common.StringPtr(""),
+				Header:                              &shared.Header{Fields: map[string][]byte{}},
+				WorkflowIdReusePolicy:               shared.WorkflowIdReusePolicyAllowDuplicateFailedOnly.Ptr(),
+				CronOverlapPolicy:                   shared.CronOverlapPolicySkipped.Ptr(),
+			},
+		},
 		{
 			name: "first run at negative",
 			options: StartWorkflowOptions{
@@ -3074,7 +3142,9 @@ func TestGetSignalWithStartRequest(t *testing.T) {
 
 			// set the randomized fields in the expected request before comparison
 			tc.wantRequest.Identity = &wc.identity
-			tc.wantRequest.RequestId = gotReq.RequestId
+			if tc.wantRequest.RequestId == nil {
+				tc.wantRequest.RequestId = gotReq.RequestId
+			}
 
 			assert.Equal(t, tc.wantRequest, gotReq)
 		})